holo_hash: reorganize hashing code to fix build failure
previous to this building with `no-default-features` was broken.
