+++ crates/holo_hash/CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## \[Unreleased\]
 
+### Fixed
+
+- Build with `no-default-features` was fixed by reorganizing hashing code.
+
 ## 0.0.5
 
 ## 0.0.4
