+++ crates/holo_hash/Cargo.toml
@@ -16,10 +16,10 @@ normal = ["tracing"]
 [dependencies]
 serde = "1"
 serde_bytes = "0.11"
+blake2b_simd = {version = "0.5.10"}
 
 arbitrary = {version = "1.0", optional = true}
 base64 = {version = "0.13", optional = true}
-blake2b_simd = {version = "0.5.10", optional = true}
 derive_more = { version = "0.99", optional = true }
 fixt = { version = "0.0.5", path = "../fixt", optional = true }
 futures = {version = "0.3", optional = true}
@@ -39,7 +39,7 @@ default = ["serialized-bytes", "fixturators"]
 full = ["fixturators", "hashing", "string-encoding", "rusqlite"]
 
 fixturators = ["fixt", "rand", "hashing", "string-encoding"]
-hashing = ["futures", "must_future", "blake2b_simd", "serialized-bytes"]
+hashing = ["futures", "must_future", "serialized-bytes"]
 serialized-bytes = ["holochain_serialized_bytes"]
-string-encoding = ["base64", "blake2b_simd", "derive_more"]
+string-encoding = ["base64", "derive_more"]
 test_utils = ["fixturators"]
