+++ crates/holo_hash/src/encode.rs
@@ -1,4 +1,5 @@
 use crate::assert_length;
+use crate::encode_common::holo_dht_location_bytes;
 use crate::error::HoloHashError;
 use crate::HashType;
 use crate::HoloHash;
@@ -95,31 +96,8 @@ pub fn holo_hash_decode(prefix: &[u8], s: &str) -> Result<Vec<u8>, HoloHashError
     Ok(s.to_vec())
 }
 
-/// internal compute the holo dht location u32
-pub fn holo_dht_location_bytes(data: &[u8]) -> Vec<u8> {
-    // Assert the data size is relatively small so we are
-    // comfortable executing this synchronously / blocking tokio thread.
-    assert_eq!(32, data.len(), "only 32 byte hashes supported");
-
-    let hash = blake2b_128(data);
-    let mut out = vec![hash[0], hash[1], hash[2], hash[3]];
-    for i in (4..16).step_by(4) {
-        out[0] ^= hash[i];
-        out[1] ^= hash[i + 1];
-        out[2] ^= hash[i + 2];
-        out[3] ^= hash[i + 3];
-    }
-    out
-}
-
 /// internal compute a 32 byte blake2b hash
 pub fn blake2b_256(data: &[u8]) -> Vec<u8> {
     let hash = blake2b_simd::Params::new().hash_length(32).hash(data);
-    hash.as_bytes().to_vec()
-}
-
-/// internal compute a 16 byte blake2b hash
-pub fn blake2b_128(data: &[u8]) -> Vec<u8> {
-    let hash = blake2b_simd::Params::new().hash_length(16).hash(data);
     hash.as_bytes().to_vec()
 }
