+++ crates/holo_hash/src/encode_common.rs
@@ -1,0 +1,24 @@
+//! Common encoding functions.
+
+/// internal compute the holo dht location u32
+pub fn holo_dht_location_bytes(data: &[u8]) -> Vec<u8> {
+    // Assert the data size is relatively small so we are
+    // comfortable executing this synchronously / blocking tokio thread.
+    assert_eq!(32, data.len(), "only 32 byte hashes supported");
+
+    let hash = blake2b_128(data);
+    let mut out = vec![hash[0], hash[1], hash[2], hash[3]];
+    for i in (4..16).step_by(4) {
+        out[0] ^= hash[i];
+        out[1] ^= hash[i + 1];
+        out[2] ^= hash[i + 2];
+        out[3] ^= hash[i + 3];
+    }
+    out
+}
+
+/// internal compute a 16 byte blake2b hash
+pub fn blake2b_128(data: &[u8]) -> Vec<u8> {
+    let hash = blake2b_simd::Params::new().hash_length(16).hash(data);
+    hash.as_bytes().to_vec()
+}
