+++ crates/holo_hash/src/lib.rs
@@ -8,11 +8,11 @@ mod has_hash;
 mod hash;
 pub mod hash_type;
 
+pub mod encode_common;
+
 pub use aliases::*;
-pub use encode::{holo_hash_decode, holo_hash_decode_unchecked, holo_hash_encode};
 pub use has_hash::HasHash;
 pub use hash::*;
-pub use hash_b64::*;
 pub use hash_type::HashType;
 pub use hash_type::PrimitiveHashType;
 
@@ -20,12 +20,21 @@ pub use hash_type::PrimitiveHashType;
 #[cfg(not(feature = "string-encoding"))]
 pub mod encode_raw;
 
+#[cfg(not(feature = "string-encoding"))]
+pub use encode_raw as encode;
+
 /// Include nice string encoding methods and From impls
 #[cfg(feature = "string-encoding")]
 pub mod encode;
 
 #[cfg(feature = "string-encoding")]
+pub use encode::{holo_hash_decode, holo_hash_decode_unchecked, holo_hash_encode};
+
+#[cfg(feature = "string-encoding")]
 pub mod hash_b64;
+
+#[cfg(feature = "string-encoding")]
+pub use hash_b64::*;
 
 #[cfg(feature = "fixturators")]
 pub mod fixt;
