+++ crates/holochain_zome_types/Cargo.toml
@@ -13,7 +13,7 @@ edition = "2018"
 
 [dependencies]
 chrono = "0.4.6"
-holo_hash = { version = "0.0.5", path = "../holo_hash", default-features = false }
+holo_hash = { version = "0.0.5", path = "../holo_hash", default-features = false, features = ["serialized-bytes"] }
 holochain_serialized_bytes = "=0.0.51"
 paste = "=1.0.5"
 serde = { version = "1.0", features = [ "derive" ] }
